+++ pkgs/espressos/src/boiler.hpp
@@ -5,7 +5,7 @@
 #include <PIDController.hpp>
 
 struct TempReading {
-  double temp;
+  float temp;
   uint8_t fault; // Fault identifier
 
   bool operator==(TempReading const &rhs) const {
@@ -50,22 +50,22 @@ class BoilerPID {
 private:
   CachedOutputPin outputPin;
   Adafruit_MAX31865 thermo;
-  double Output;
-  double Kp = BOILER_PID_P, Ki = BOILER_PID_I, Kd = BOILER_PID_D;
+  float Output;
+  float Kp = BOILER_PID_P, Ki = BOILER_PID_I, Kd = BOILER_PID_D;
   int WindowSize = 100;
   unsigned long windowStartTime;
-  PIDController<double, double, unsigned long> pid;
+  PIDController<float, float, unsigned long> pid;
   TempReading temp;
 
 public:
   BoilerPID(int relayPin, int max31865SPIPin, SPIClass *theSPI);
 
   void setup(unsigned long now);
-  void SetSetPoint(double setPoint);
+  void SetSetPoint(float setPoint);
 
   struct TempReading getTemp();
 
-  void SetTunings(double Kp, double Ki, double Kd);
+  void SetTunings(float Kp, float Ki, float Kd);
 
   bool loop(unsigned long now);
 };
